video-engine/requirements.txt
# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Image/Video Processing
Pillow>=10.0.0
opencv-python>=4.8.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
numpy>=1.24.0

# AI/ML Models (optional - for local generation)
# torch>=2.0.0
# diffusers>=0.21.0
# transformers>=4.33.0
# accelerate>=0.23.0

# API Clients
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Background Jobs
celery>=5.3.0
redis>=5.0.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
tenacity>=8.2.0
